Summary of the website
This document contains all the screens contained within the website (and a short description of whats going on). I tried my best to upload the entire website onto a free public server, but there was just to many limitations and problems (such max querying). The website runs on the QUT's private server and a VPN with my credentials is needed to gain access.